+++ src/pages/Junior.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Junior() {
+    return <h1>Junior</h1>
+}
+
+export default Junior
