+++ src/pages/JuniorPlus.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function JuniorPlus() {
+    return <h1>JuniorPlus</h1>
+}
+
+export default JuniorPlus
